requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
fastapi==0.105.0
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
